requirements.txt
defusedxml==0.4.1
dj-database-url==0.4.0
Django==1.9.4
django-allauth==0.25.2
django-appconf==1.0.1
django-compressor==2.0
django-social-auth==0.7.28
django-static-precompiler==1.3
gunicorn==19.4.5
httplib2==0.9.2
oauth2==1.9.0.post1
oauthlib==1.0.3
psycopg2==2.6.1
python-openid==2.2.5
python3-openid==3.0.10
PyYAML==3.11
rcssmin==1.0.6
requests==2.9.1
requests-oauthlib==0.6.1
rjsmin==1.0.12
simplejson==3.8.2
six==1.10.0
virtualenv==15.0.1
whitenoise==3.0
